requirements.txt
fastapi
pydantic
pypdf
python-docx
python-pptx
phonenumbers
PyMuPDF